Make this package installable for python
