+++ setup.py
@@ -1,0 +1,10 @@
+from setuptools import setup, find_packages
+setup(
+    name="JSON_minify",
+    version="0.2",
+    author='Gerald Storer',
+    packages=find_packages(),
+    install_requires=[],
+    zip_safe=False,
+    entry_points={}
+)
